Enabled warnings for brace style and made changes as needed.
Most of the code already used the default brace style, commonly refered to in
JavaScript as "one true brace style". Braces go on the same line as the if,
else, while, etc. keywords:

if(expression) {
} else {
}

Violations of the rule will result in a warning, so it won't block the tests
from succeeding.